Added a base for the entity component system

+++ Assets/Scripts/Entities/Stats/StatModifier.cs
@@ -5,7 +5,7 @@ namespace Entities
         public float Value { get; private set; }
         public string Source { get; private set; }
 
-        public StatModifier(int value, string source)
+        public StatModifier(float value, string source)
         {
             this.Value = value;
             this.Source = source;
